Allows differential exp plot to pick covariates
These are the companion changes to the ones I made in server.R to the "differential expression" boxplot. Mostly, it allows the inputs to be from the phyloseq object instead of "batch" and "condition" vectors, and it allows users to select covariates of interest from a list. I added a copy of findPhyseqData() in order to populate the pull-down menu for covariates, but I don't expect that to be the final solution here.